Conflicts:
	src/main/java/se/kth/bbc/study/StudyMB.java
	src/main/java/se/kth/kthfsdashboard/user/RegistrationController.java

+++ src/main/java/se/kth/bbc/study/StudyController.java
@@ -53,12 +53,12 @@ public class StudyController {
     }
     
     
-    public long getAllStudy(String username){
-        return (Long)em.createNamedQuery("TrackStudy.findAllStudy").setParameter("username", username).getSingleResult();
+    public int getAllStudy(String username){
+        return ((Long)em.createNamedQuery("TrackStudy.findAllStudy").setParameter("username", username).getSingleResult()).intValue();
     }
     
-    public long getMembers(String name){
-        return (Long)em.createNamedQuery("TrackStudy.findMembers").setParameter("name", name).getSingleResult();
+    public int getMembers(String name){
+        return ((Long)em.createNamedQuery("TrackStudy.findMembers").setParameter("name", name).getSingleResult()).intValue();
     }
     
     public List<TrackStudy> filterPersonalStudy(String username){
